+++ packages/expo-audio/src/utils/resolveSource.ts
@@ -1,5 +1,5 @@
 import { AudioSource } from '../Audio.types';
-import resolveAssetSource from '../resolveAssetSource';
+import resolveAssetSource from './resolveAssetSource';
 
 export function resolveSource(source?: AudioSource | string | number | null): AudioSource | null {
   if (typeof source === 'string') {
